+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 * **[Feat]**: added support to disable Gutenberg Editor
 * **[Refactor]**: removed public router() method from App class
 * **[Refactor]**: refactored app() function
+* **[Refactor]**: refactored QueryBuilders and related models
+* **[Refactor]**: refactored console command root folder as protected parent property
 * **[Fix]**: improved model stub for Post Type and Taxonomy
+* **[Fix]**: removed Product post type from protected - let user handle which to register
 
-Release Date: August 17, 2021
+Release Date: August 18, 2021
 
 ## Core Release